Fix enter key check while composing characters
Chrome on macOS sends a KeyEvent with `key` set to `Enter`, the keyCode `229` and `isComposing` set to `true` when pressing enter while a character composition sequence is active, then an event for the composed key and ends with a Enter key event with the correct key code.
This has thrown our Enter key detection off and resulted in sending of the wanted message and a public message with the single composed character as content.
Ignoring the Enter key while characters are being composed seems to be safe.